Add dwave.embedding.zephyr module
This module defines functions for finding cliques and bicliques in
Zephyr-structured graphs.

Note the analogous modules in dwave.embedding, chimera and pegasus, use the
minorminer.utils package to provide this functionality.  This module uses the
minorminer.busclique package instead, which is faster and supports all three
topologies.  In the future, the chimera and pegasus modules could be updated
to used minorminer.busclique as well, and the functions defined in this commit
could be used to support all three topologies with a small refactoring.